i think it's last pull

+++ app/controllers/operator/users_controller.rb
@@ -15,6 +15,7 @@ module Operator
       @cashier_users = User.where(role_id: 5).order(:name).page(params[:cashier_users].to_i)
       @user_admin_users = User.where(role_id: 3).order(:name).page(params[:user_admin_users].to_i)
       @credit_admin_users = User.where(role_id: 4).order(:name).page(params[:credit_admin_users].to_i)
+      @message = flash[:message] if flash[:message] != nil
     end
 
     def new
@@ -23,9 +24,33 @@ module Operator
     end
 
     def create
+
+      #создание сотрудника банка
+      employee_name = params[:user][:man_name]
+      employee_surname = params[:user][:man_surname]
+      employee_patronymic = params[:user][:man_patronymic]
+      employee_mobile_phone = params[:user][:man_mobile_phone]
+      employee_email = params[:user][:man_email]
+      bank_employee = BankEmployee.create_employee_for_user(employee_name,
+                                                            employee_surname,
+                                                            employee_patronymic,
+                                                            employee_email,
+                                                            employee_mobile_phone)
+      # # создание пользователя
       @user = User.new(user_params)
-      @user.save_first_time
-      @user_inputs = {name: @user.name, role_id: @user.role_id}
+      @user = @user.save_user_employee(bank_employee)
+      # puts json: user
+      # #обновление связи user и bank_employee
+      bank_employee.update_attributes(user_id: @user.id)
+
+      @user_inputs = {name: @user.name,
+                      role_id: @user.role_id,
+                      man_name: employee_name,
+                      man_surname: employee_surname,
+                      man_patronymic: employee_patronymic,
+                      man_mobile_phone: employee_mobile_phone,
+                      man_email: employee_email}
+
       redirect_to :back, flash: {validation_errors: @user.error_message,
                                  inputs_params: params[:user]} unless @user.error_message.nil?
     end
@@ -38,35 +63,30 @@ module Operator
       @user = User.find(params[:id])
       @user.generate_password
       @user.save
-      @user.send_email("Пароль был изменен.\nПользователь\n имя: #{@user.name}\n новый пароль: #{@user.password}")
-      @user.send_sms("Пароль был изменен.\nПользователь\n имя: #{@user.name}\n новый пароль: #{@user.password}")
+      @user.send_email("Пароль был изменен. Логин: #{@user.name}, новый пароль: #{@user.password}.")
+      @user.send_sms("Пароль был изменен. Логин: #{@user.name}, новый пароль: #{@user.password}.")
     end
     def update
       @user.update(user_params)
 
-      @client_users = User.where(role_id: 1).order(:name).page(params[:client_users_page].to_i)
-      @operator_users = User.where(role_id: 2).order(:name).page(params[:operator_users_page].to_i)
-      @cashier_users = User.where(role_id: 5).order(:name).page(params[:cashier_users].to_i)
-      @user_admin_users = User.where(role_id: 3).order(:name).page(params[:user_admin_users].to_i)
-      @credit_admin_users = User.where(role_id: 4).order(:name).page(params[:credit_admin_users].to_i)
       @message = 'Пользователь обновлен.'
-      render 'operator/users/index'
+
+      redirect_to operator_users_path, flash: { message: @message }
     end
 
     def destroy
       begin
         user = User.find(params[:id])
-        user.send_email("Пользователь с именем #{user.name} был удален")
-        user.send_sms("Пользователь с именем #{user.name} был удален")
-        user.destroy
-
-        @client_users = User.where(role_id: 1).order(:name).page(params[:client_users_page].to_i)
-        @operator_users = User.where(role_id: 2).order(:name).page(params[:operator_users_page].to_i)
-        @cashier_users = User.where(role_id: 5).order(:name).page(params[:cashier_users].to_i)
-        @user_admin_users = User.where(role_id: 3).order(:name).page(params[:user_admin_users].to_i)
-        @credit_admin_users = User.where(role_id: 4).order(:name).page(params[:credit_admin_users].to_i)
         @message = 'Пользователь удален.'
-        render 'operator/users/index'
+        if user.id != current_user.id
+          user.send_email("Пользователь с именем #{user.name} был удален.")
+          user.send_sms("Пользователь с именем #{user.name} был удален.")
+          user.destroy
+        else
+          @message = 'Вы не может удалить себя.'
+        end
+
+        redirect_to operator_users_path, flash: { message: @message }
       end
     end
